[UPDATE] WebClientResponseException 처리 로직
기존 방식 : 429 에러 발생 시 컨슈머 스레드를 sleep => 해당 스레드가 sleep된 상태에서 리밸런싱 시 컨슈머 그룹에 속하지 못하고 이탈되며 해당 스레드가 running될 때 다시 컨슈머 poll을 호출하여 리밸런싱 발생

최신 방식 : 429 에러 발생 시 컨슈머를 pause 하여 poll() 시 레코드를 반환하지 않도록 만듬. 동시에 2분 후 해당 컨슈머를 resume하는 로직 추가 -> pause된 컨슈머도 리밸런스 발생 시 컨슈머 그룹에 응답하여 리밸런싱에 참여 => resume 될 때 다시 리벨런싱 발생 x

+++ src/main/java/com/lolsearcher/persistance/failmatchids/constant/RiotGamesConstants.java
@@ -4,5 +4,5 @@ public class RiotGamesConstants {
 
     public static final String ASIA_BASE_URL = "https://asia.api.riotgames.com";
 
-    public static final int REQUEST_WAIT_MS = 120000; /* 2min */
+    public static final long REQUEST_WAIT_MS = 120000; /* 2min */
 }
